ai-service/requirements.txt
flask==2.3.3
gunicorn==21.2.0
openai==1.3.5
python-dotenv==1.0.0
tiktoken==0.5.1
requests==2.31.0
numpy==1.26.0
pandas==2.1.1
transformers==4.35.0
scikit-learn==1.3.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
